wip brew hook

+++ src/file.ts
@@ -22,6 +22,7 @@ export async function rename(from: string, to: string) {
 }
 
 export async function remove(file: string) {
+  if (!await exists(file)) return
   debug('remove', file)
   return deps.fs.remove(file)
 }
@@ -54,4 +55,8 @@ export async function readJSON(file: string) {
 export async function outputJSON(file: string, data: any, options: FS.WriteOptions = {}) {
   debug('outputJSON', file)
   return deps.fs.outputJSON(file, data, { spaces: 2, ...options })
+}
+
+export function realpathSync(p: string) {
+  return deps.fs.realpathSync(p)
 }
